refactor : chat 관련 엔티티 패키지 - ChatRoom - ChatMessage
